Add editing of existing surveys
Only a limited set of operations can be done to
avoid ending up in inconsistent states (or make
the solution for it very complex)

The "backend" implementation of the edit operations
could have been done a bit cleaner (I think) if I had
prepared better access rules. But I didn't want to
change that now, so this will have to do.