Prevent adding responders to ended surveys
In case the survey has ended, one way or the other, reviving it
by adding additional responders must also imply extending
the open hours.

This also fixes an issue where an "manually" completed survey
became a canceled survey just because more responders were
later on added. (This was because the hoursOpen has 0 in that
case).