Team-ID related improvement and robustifications
TeamAdmin now shows a brief summary of the selected team.
While working with this I realized that different states will
update such that the loaded surveys will not always correspond
to the current team selection (due to update delays).
To have better control of this the "applicble" team-ID is now
also returned from the useTeamTracker hook.