Tweak QueryModal to support cascaded queries
The queryConfirm/QueryModal design is not designed for
concurrent queries but I think it should support launching
of a second query in the callback from a first one.
I think this solves that but I might have missed some race
condition w.r.t. the modal's inner workings (inkl animations)