Add ControlledSelect and RouteSelect components.
These will eventuall be used in a new design for how team
selection is managed. Hopefully in a less convoluted way
while still enabling  a set of "complex scenarios"

For now the user input is managed via a native "select" element.
Don't know if this what I want in the end... very hard to make up
my mind about that!