Avoid "dummy" md-file load for survey results
In some situations, e.g. when manually ticking the
"completed" box for an ongoing survey, the analysis
part is null for a brief period due to the interim
"canceled" state. (Why? can't remember now...
maybe that should shange)

Anyway, due to the default fallback in ResultInterpretation
that lead to a dummy fetch of the "no-match" content
before the actual match was derived.