First attempt at lazy-loading
I'm not sure about the Start page; for now it is lazy-loaded
(as very few actual users visits it).

Also, corrected bug if aborting a slow load of a md-page.
(I have a feeling this was not the only "leak" of this kind...
I'l have to review the rest of the code at some point)