Fix auth-related issues seen on mobile
After adding the CSP headers, eveything looked "OK" for  a
while. But on mobile, after a few hours of inactivity after
login, access to Firebase stopped working.

Due to to struggels of obtaining the console log on mobile
the reason was not obvious but it turned out that an attempt
to connect to https://securetoken.googleapis.com was blocked.

(Why only on mobile?)

So, relax the CSP restrictions a bit by allowing all googleapis-
domains.

(I really have to integrate Sentry or similar to detect these
things in an easier way)