Major rework of SurveyResults and friends
This is still work in progress but now a new
EfficiencyResult is also presented in addition to
the MaturityResult.

This also required a rework of the MaturityResult
component, including its styling.

For now there is a lot of overlap/duplication
between the MaturityResult and EfficiencyResult
styling, I should probably try to do something
about that.

As a consequence of these changes, I had to break
out the "details part" from ResultInterpretation
(presented in different parts of SurveyResults now)

Note that the history trend plot is still not updated
(doesn't fit at all right now, must fix that...)