First step towards new "v_users" access lest
Introducing a list of "validated users" in the database.
Eventually, this list will control access (instead of x_users)
and different strategies for "validation" can be applied.
E.g. email-verification might be required. Or maybe even
manual intervention (gating).
But for now, the list is unconditionally populated for
any user who manages to log in.