Avoid janky width during MarkdownBlock transitions
Avoid short period of height=0 while not showing any content, i.e.
before the new content has been loaded. This avoids short periods
of hidden scroll bar which would otherwise affect the width of the
layout area.

Because of this change I had to revise some details around how
to actually ensure that a new MarkdownPage load implies a scroll
to the top.