Keep track of "owned teams" in AppContext
This means that CreatorApp and Results don't have to acquire
this information every time they mount. This prevents unnecessary
re-subscriptions every time the user navigates back and forth
between Results and the Creator pages.

Also, in case of error accessing the team-information, a logout is
triggered. (Not a "solution") but the Creator parts are quite useless
without this info)